Specify prompt kwarg to all model calls
Some models may not support the prompt as a positional and only allow it as a kwarg.

The main case for this is when you are trying to inherit from `DeepEvalBaseLLM`. It is not a valid override to add a positional `prompt` argument as `DeepEvalBaseLLM` does not have it, yet DeepTeam is assuming that the first positional argument is present. This is making every generate call inherently unsafe and not interoperable with DeepEval models.